Refactor: 주문 취소 가능한지 MemberContext로 권한 검증, 주문 취소 상태 확인 기능 추가

+++ 4Week_Misson/src/main/java/com/ll/exam/sen_books/app/order/controller/OrderController.java
@@ -27,6 +27,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.client.ResponseErrorHandler;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.annotation.PostConstruct;
 import java.util.Base64;
@@ -85,9 +86,13 @@ public class OrderController {
 
     @GetMapping("/{id}/cancel")
     @PreAuthorize("isAuthenticated()")
-    public String removeOrder(@PathVariable Long id) {
+    public String removeOrder(@AuthenticationPrincipal MemberContext memberContext, @PathVariable Long id) {
         Order order = orderService.findById(id).get();
+        Member member = memberContext.getMember();
 
+        if(orderService.canCancel(member, order) == false) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         orderService.removeOrder(order);
 
         return "redirect:/order/list?msg=" + Ut.url.encode("%d번 품목을 삭제하였습니다.".formatted(id));
